api-sam3/requirements.txt
fastapi
uvicorn
python-multipart
transformers>=4.48.0
torch
torchvision
Pillow
requests
huggingface_hub
